backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlalchemy>=2.0.0
pydantic>=2.0.0
sse-starlette>=2.1.0
pymupdf>=1.24.0
python-docx>=1.1.0
sentence-transformers>=3.0.0
scikit-learn>=1.5.0
openai>=1.40.0
google-genai>=1.0.0
python-multipart>=0.0.9
openpyxl>=3.1.0
